+++ tests/unit-tests/Ares/Rule/MaxRuleTest.php
@@ -47,6 +47,7 @@ class MaxRuleTest extends TestCase
      *
      * @testWith ["float"]
      *           ["integer"]
+     *           ["numeric"]
      *
      * @param string $type Supported type.
      * @return void