+++ tests/unit-tests/Ares/Schema/TypeTest.php
@@ -27,6 +27,7 @@ class TypeTest extends TestCase
      *           ["INTEGER", "integer"]
      *           ["LIST", "list"]
      *           ["MAP", "map"]
+     *           ["NUMERIC", "numeric"]
      *           ["STRING", "string"]
      *           ["TUPLE", "tuple"]
      *